Deprecated BF16 support in DepthConvert
* Removed BF16 validation tests for DepthConvert

* Revert back to using inline assembly to convert to/from BF16

* Resolves COMPMID-5800

Reviewed-on: https://review.mlplatform.org/c/ml/ComputeLibrary/+/8929
Comments-Addressed: Arm Jenkins <bsgcomp@arm.com>
Benchmark: Arm Jenkins <bsgcomp@arm.com>
Tested-by: Arm Jenkins <bsgcomp@arm.com>

+++ utils/TypePrinter.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017-2022 Arm Limited.
+ * Copyright (c) 2017-2023 Arm Limited.
  *
  * SPDX-License-Identifier: MIT
  *
@@ -482,7 +482,7 @@ inline ::std::ostream &operator<<(::std::ostream &os, const BoundingBoxTransform
 inline ::std::ostream &operator<<(::std::ostream &os, const bfloat16 &v)
 {
     std::stringstream str;
-    str << static_cast<float>(v);
+    str << v;
     os << str.str();
     return os;
 }
